+++ src/application/services/folders/folders.service.ts
@@ -1,3 +1,4 @@
+import { RegisteredFolderErrorStatus, RegisteredFolderStatus } from 'src/domain/entities/RegisteredFolderStatus';
 import { RomachEntitiesApiInterface } from 'src/application/interfaces/romach-entities-api.interface';
 import { RomachRepositoryInterface } from 'src/application/interfaces/romach-repository.interface';
 import { BasicFolderChange } from 'src/application/interfaces/basic-folder-changes.interface';
@@ -14,7 +15,120 @@ export class FoldersService {
         private readonly repository: RomachRepositoryInterface,
     ) {}
 
-    updateFoldersToRegisteredFolders(registeredFolders: RegisteredFolder[], folders: Folder[]) {        
+    async upsertGeneralerror(
+        upn: string,
+        folderId: string,
+        isPasswordProtected: boolean,
+    ): Promise<Result<Folder | void, RegisteredFolderErrorStatus>> {
+        const input = {
+            upn,
+            folderId,
+            isPasswordProtected,
+            lastValidPasswordTimestamp: Timestamp.ts1970(), // need to think about it
+        };
+        const newRegisteredFolderResult = RegisteredFolder.createGeneralErrorRegisteredFolder(input);
+        if (newRegisteredFolderResult.isFail()) {
+            this.logger.error('failed to create new registeredFolder');
+            return Result.fail('general-error');
+        }
+
+        const newRegisteredFolder = newRegisteredFolderResult.value();
+        const upsertFolderResult = await this.repository.upsertRegisteredFolder(newRegisteredFolder);
+        if (upsertFolderResult.isFail()) {
+            this.logger.error('failed to upsert registeredFolder to repo');
+            return Result.fail('general-error');
+        }
+        return Result.fail();
+    }
+
+    async upsertValid(upn: string, folderId: string, folder: Folder, password?: string): Promise<Result<Folder>> {
+        const registeredFoldersResult = await this.repository.getRegisteredFoldersByIdAndPassword(folderId, password);
+        const currentregisteredFolders = registeredFoldersResult.value();
+        if (!currentregisteredFolders)
+            return Result.fail('failed to get registeredFolders with same folderId and password');
+
+        const changedValidregisteredFoldersResult = this.updateFolderToRegisteredFolders(
+            currentregisteredFolders,
+            folder,
+        );
+        if (changedValidregisteredFoldersResult.isFail()) {
+            this.logger.error('failed to update registeredFolders');
+            return Result.fail();
+        }
+
+        const newregisteredFolderResult = RegisteredFolder.createValidRegisteredFolder({
+            upn,
+            folder,
+            password,
+            lastValidPasswordTimestamp: Timestamp.now(),
+        });
+        if (newregisteredFolderResult.isFail()) {
+            this.logger.error('failed to create new registeredFolder');
+            return Result.fail();
+        }
+
+        const newregisteredFolder = newregisteredFolderResult.value();
+        const updatedRegisteredFolders = changedValidregisteredFoldersResult.value();
+        if (updatedRegisteredFolders) {
+            const upsertFolderResult = await this.repository.upsertRegisteredFolders([
+                newregisteredFolder,
+                ...updatedRegisteredFolders,
+            ]);
+            if (upsertFolderResult.isFail()) {
+                this.logger.error('failed to upsert registeredFolder to repo');
+                return Result.fail();
+            }
+        }
+
+        return Result.Ok(folder);
+    }
+
+    async upsertWrongPassword(upn: string, folderId: string) {
+        const currentregisteredFoldersResult = await this.repository.getRegisteredFoldersById(folderId);
+        const currentregisteredfolders = currentregisteredFoldersResult.value();
+        if (!currentregisteredfolders) return Result.fail(currentregisteredFoldersResult.error()); //registeredFoldersResult ?? - i could heap up logs
+
+        const newregisteredFolderResult = RegisteredFolder.createWrongPasswordRegisteredFolder({
+            upn,
+            folderId,
+        });
+        const newregisteredFolder = newregisteredFolderResult.value();
+        if (!newregisteredFolder) return Result.fail('failed to create new registeredFolder');
+
+        const changedregisteredFoldersResult = this.changeStatusToregisteredFolders(
+            currentregisteredfolders,
+            'wrong-password',
+        );
+        const changedregisteredFolders = changedregisteredFoldersResult.value();
+        if (!changedregisteredFolders) return Result.fail(changedregisteredFoldersResult.error());
+
+        const allregisteredFoldersToUpsert = [...changedregisteredFolders, newregisteredFolder];
+        const upsertFolderResult = await this.repository.upsertRegisteredFolders(allregisteredFoldersToUpsert);
+        if (upsertFolderResult.isFail()) {
+            this.logger.error('failed to upsert registeredFolder to repo');
+            return Result.fail();
+        }
+    }
+
+    changeStatusToregisteredFolders(registeredFolders: RegisteredFolder[], newStatus: RegisteredFolderStatus) {
+        const createregisteredFolder = RegisteredFolder.getCreateFunctionByStatus(newStatus);
+
+        const createregisteredfoldersResult = registeredFolders.map((registeredFolder) =>
+            createregisteredFolder({
+                ...registeredFolder.getProps(),
+                lastValidPasswordTimestamp: newStatus === 'valid' ? Timestamp.now() : null,
+            }),
+        );
+
+        if (Result.combine(createregisteredfoldersResult).isFail()) {
+            this.logger.error('failed to change status to registeredFolders');
+            return Result.fail();
+        }
+
+        return Result.Ok(createregisteredfoldersResult.map((x) => x.value()));
+    }
+
+    updateFoldersToRegisteredFolders(registeredFolders: RegisteredFolder[], folders: Folder[]) {
         const yair = folders.flatMap((folder) => {
             const registeredFoldersWithSameFolder = filter(
                 registeredFolders,
@@ -37,7 +151,9 @@ export class FoldersService {
 
     updateFolderToRegisteredFolders(registeredFolders: RegisteredFolder[], folder: Folder) {
         const createregisteredfoldersResult = registeredFolders.map((registeredFolder) => {
-            const createregisteredFolder = RegisteredFolder.getCreateFunctionByStatus(registeredFolder.getProps().status);
+            const createregisteredFolder = RegisteredFolder.getCreateFunctionByStatus(
+                registeredFolder.getProps().status,
+            );
             return createregisteredFolder({
                 ...registeredFolder.getProps(),
                 folder,
@@ -64,8 +180,9 @@ export class FoldersService {
     async userMutationFolderInterval(upn: string, folderIds: string[]) {
         const getRegisteredFoldersByUpnResult = await this.repository.getRegisteredFoldersByUpn(upn);
         const registeredFolders = getRegisteredFoldersByUpnResult.value();
-        const [relevantregisteredFolders, irrelevantregisteredFolders] = partition(registeredFolders, (registeredFolder) =>
-            folderIds.includes(registeredFolder.getProps().folderId),
+        const [relevantregisteredFolders, irrelevantregisteredFolders] = partition(
+            registeredFolders,
+            (registeredFolder) => folderIds.includes(registeredFolder.getProps().folderId),
         );
 
         const irrelevantregisteredFoldersIds = irrelevantregisteredFolders.map(
@@ -123,8 +240,6 @@ export class FoldersService {
                 where registration_timestamp > 60s or valid_password_timestamp > 24h
     */
 
-
-                
 // private handleUpdatedBasicFoldersReactive(
 //     updatedBasicFolders: BasicFolder[]
 // ): Observable<Result<void>> {
